This PR address #8408 (#8586)
It adds description to the `Ingress Percentile Response Times and Transfer Rates` view, so that user knows that this latency data is independant of dashboard time range.

This PR also adds two new panel about the latency, where one shows latency as a timeseries graph and other shows heatmap of the latency distribution.